Rework line-tracking turn logic and add gyro-assisted turn detection
Replace the weighted-average+PID steering with a threshold-based state
machine (straight / approach / pivot-turn) so the car only reacts to
multi-sensor deviations instead of every minor wobble, with a light
proportional correction for small drift while going straight.

Add a lightweight MPU6050 driver (raw gyro Z-axis integration, no DMP)
over I2C0 so 90-degree turns are measured against actual yaw rotation
instead of guessed from grayscale sensor patterns, with the sensor
pattern check kept as a fallback exit condition.

Also fixes the TB6612 motor direction pins (AIN1/AIN2 swapped) that
caused the car to drive backwards.

+++ empty_LP_MSPM0G3507_nortos_ticlang/Hardware/xunji.h
@@ -11,10 +11,10 @@ extern volatile uint8_t Gray_Data[8];
 //巡线闭环给 Velocity_A/Velocity_B 的目标转速(RPM)，由 Line_Track_Task 每次调用后刷新
 extern int Target_RPM_A, Target_RPM_B;
 
-//当前巡线偏差，调试/OLED显示可以直接读
-extern int8_t Track_Err;
+//当前左右触线数量差(右-左)，调试/OLED显示可以直接读，不参与控制
+extern float Track_Err;
 
-//在main()的while(1)里循环调用；内部会阻塞扫描8路灰度(约8ms)，遇到急转弯分支会阻塞多等250~300ms
+//在main()的while(1)里循环调用；内部会阻塞扫描8路灰度(约8ms)，不再有额外的原地转弯阻塞延时
 void Line_Track_Task(void);
 
 #endif
